refactor: extract UI helper functions to src/ui/helpers.py
- Created src/ui/helpers.py with UI utility functions:
  - load_actions_config()
  - load_llms_file() / save_llms_file()
  - discover_and_preview_llmstxt()
  - import_llmstxt_actions()
  - preview_merge_llmstxt()
  - load_env_browser_settings_file()
- Reduced bykilt.py from 3198 to 2988 lines (210 lines moved)
- Verified import successful

Issue: #326 (Phase 3.1)
